Move validateAndFixWindowPosition() out of window config. Don't calculate valid location, use default
